Add maxresplen for clarity (wrt SSL_read)
This addresses clang warnings about implicit casts from (size_t) to (int),
arising due to the previous
	SSL_read(..., *resplen)
since SSL_read takes the maximum length as an (int).

We could silence those warnings with a smaller diff by using
(int)(*resplen), but I think the code is clearer if we add this new
variable.